Dodati kontroleri

+++ SalonNamestaja/DTOs/PorudzbinaDto.cs
@@ -5,7 +5,7 @@
         public int PorudzbinaID { get; set; }
         public DateTime DatumVreme { get; set; }
         public string Status { get; set; }
-        public decimal UkupanIznos { get; set; }
+        public double UkupanIznos { get; set; }
 
         public int KupacID { get; set; }
         public int ProdavacID { get; set; }
